modularization full of codes and structures updated.
